src/requirements.txt
Django==4.1
psycopg2-binary==2.9.9
django-cors-headers==4.2.0
celery==5.2.7
gunicorn==20.0.4
python-telegram-bot==13.6
django-environ==0.10.0
djangorestframework==3.14.0
django-extensions==3.2.1
django-silk==5.0.3
requests==2.30.0
django-debug-toolbar==4.0.0
whitenoise==6.4.0
flake8==6.1.0
gevent==23.7.0
geopy==2.4.1
